web/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
clickhouse-connect
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.115.5
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
griffe==1.14.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
lz4==4.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.9.0
openai-agents==0.6.2
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.47.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3>=2.0.0
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0